pulled createFoodInMeal function into domIndex.js
