+++ lib/domIndex.js
@@ -1,4 +1,4 @@
-import { deleteMealFood } from './index'
+import { deleteMealFood, postFoodToMeal } from './index'
 
 const mealCalories = {
   "breakfast": 400,
@@ -114,4 +114,16 @@ export const removeMealFood = (event) => {
   let mealId = event.target.parentNode.parentNode.parentNode.parentNode.className.slice(-1)
   $tr.remove()
   deleteMealFood($tr.className, mealId)
+}
+
+export const createFoodInMeal = (event) => {
+  event.preventDefault()
+  let mealId = event.target.id.slice(-1)
+  $('#all-foods-diary tr').each((index, row) => {
+    if (index != 0 && $(row).find(':checkbox:checked').length === 1) {
+      let foodId = parseInt(row.id)
+      postFoodToMeal(mealId, foodId)
+    }
+  })
+  $('input[type=checkbox]').prop('checked', false)
 }
