+++ lib/index.js
@@ -1,4 +1,4 @@
-import { appendMeals, getEachFoodDiary, filterFoods, removeMealFood } from './domIndex'
+import { appendMeals, getEachFoodDiary, filterFoods, removeMealFood, createFoodInMeal } from './domIndex'
 import { validateForm, getEachFood, removeFood, removeValidations, createInputField, removeInputField } from './domFoods'
 import { postFood, getAllFoodsValidation, createFood } from './serviceFoods'
 import { getAllFoodsDiary, countIncrement, getAllMeals } from './serviceIndex'
@@ -19,18 +19,6 @@ export const handleResponse = (response) => {
   })
 }
 
-const createFoodInMeal = (event) => {
-  event.preventDefault()
-  let mealId = event.target.id.slice(-1)
-  $('#all-foods-diary tr').each((index, row) => {
-    if (index != 0 && $(row).find(':checkbox:checked').length === 1) {
-      let foodId = parseInt(row.id)
-      postFoodToMeal(mealId, foodId)
-    }
-  })
-  $('input[type=checkbox]').prop('checked', false)
-}
-
 const postMealConfig = () => {
   return {
     method: 'POST',
@@ -38,7 +26,7 @@ const postMealConfig = () => {
   }
 }
 
-const postFoodToMeal = (mealId, foodId) => {
+export const postFoodToMeal = (mealId, foodId) => {
   fetch(`http://quantifiedself-backend.herokuapp.com/api/v1/meals/${mealId}/foods/${foodId}`, postMealConfig())
     .then(() => getAllMeals())
     .catch((error) => console.error({error}))
